refactor: Promote E4S theme internal helpers into lib-level helpers and export them
