+++ src/components/AppBarNavigation/RenderLogo/styles.ts
@@ -9,7 +9,7 @@
 
 /* eslint-disable import/prefer-default-export */
 
-import { styled } from '@actinc/dls/helpers/material/styled';
+import { styled } from '@actinc/dls/helpers/styled';
 
 export const StyledImg = styled('img')<{ $isWrapped: boolean }>(
   ({ $isWrapped }) => ({
