+++ src/components/DataTable/TableCellBody/styles.ts
@@ -12,7 +12,7 @@
 import { TableCell } from '@mui/material';
 import { tableCellClasses } from '@mui/material/TableCell';
 
-import { styled } from '@actinc/dls/helpers/material/styled';
+import { styled } from '@actinc/dls/helpers/styled';
 
 export const StyledTableCell = styled(TableCell)(({ theme }) => ({
   [`&.${tableCellClasses.root}`]: {
