+++ src/components/SearchBar/styles.ts
@@ -14,7 +14,7 @@ import IconButton, {
   iconButtonClasses,
 } from '@mui/material/IconButton';
 
-import { styled } from '@actinc/dls/helpers/material/styled';
+import { styled } from '@actinc/dls/helpers/styled';
 
 export const StyledIconButton = styled(IconButton)<
   IconButtonProps & {
