+++ src/helpers/sort.ts
@@ -14,7 +14,10 @@ import { SortObject } from '@actinc/dls/types';
 
 type SortFunction<T> = (itemA: T, itemB: T) => number;
 
-function sort<T>({ sortBy, sortDirection }: SortObject): SortFunction<T> {
+export function sort<T>({
+  sortBy,
+  sortDirection,
+}: SortObject): SortFunction<T> {
   return (
     { [sortBy as keyof T]: itemA },
     { [sortBy as keyof T]: itemB },
