+++ src/styles/themeActEt/components.ts
@@ -63,7 +63,11 @@ export const components: Components = {
   },
   MuiAppBar: {
     defaultProps: {},
-    styleOverrides: {},
+    styleOverrides: {
+      root: {
+        height: DIMS.CUSTOM.heights.header,
+      },
+    },
   },
   MuiAutocomplete: {
     defaultProps: {},
