--- src/styles/themeEncouraDatalab/props.tsx
@@ -1,56 +1,0 @@
-/**
- * Copyright (c) ACT, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * @prettier
- */
-
-import ChevronDown from '@actinc/dls/icons/ChevronDown';
-import FilterVariant from '@actinc/dls/icons/FilterVariant';
-import { ComponentsProps } from '@mui/material/styles';
-
-const PROPS: ComponentsProps = {
-  MuiAlert: {
-    closeText: 'Close',
-    severity: 'info',
-    variant: 'standard',
-  },
-  MuiButton: {
-    disableElevation: true,
-    variant: 'contained',
-  },
-  MuiButtonGroup: {
-    disableElevation: true,
-  },
-  MuiDataGrid: {
-    components: {
-      ColumnFilteredIcon: FilterVariant,
-    },
-    disableRowSelectionOnClick: true,
-  },
-  MuiPaper: {
-    elevation: 3,
-  },
-  MuiSelect: {
-    IconComponent: ChevronDown,
-    variant: 'outlined',
-  },
-  MuiTextField: {
-    color: 'secondary',
-    InputLabelProps: {
-      shrink: true,
-    },
-    InputProps: {
-      notched: false,
-    },
-    size: 'small',
-    variant: 'outlined',
-  },
-  MuiTooltip: {
-    arrow: true,
-  },
-};
-
-export default PROPS;
